--- a/src/prng/isaac_array.rs
+++ b/src/prng/isaac_array.rs
@@ -14,6 +14,9 @@
 // implement `AsRef`, `Default`, `Serialize`, `Deserialize`, or any other
 // traits for that matter.
 
+use core::slice;
+use core::convert::AsRef;
+use core::default::Default;
 #[cfg(feature="serde1")] use serde::{Serialize, Deserialize};
 
 const RAND_SIZE_LEN: usize = 8;
@@ -23,39 +26,65 @@ const RAND_SIZE: usize = 1 << RAND_SIZE_LEN;
 #[derive(Copy, Clone)]
 #[allow(missing_debug_implementations)]
 #[cfg_attr(feature="serde1", derive(Serialize, Deserialize))]
-pub struct IsaacArray<T> {
+pub struct IsaacArray32 {
     #[cfg_attr(feature="serde1",serde(with="isaac_array_serde"))]
     #[cfg_attr(feature="serde1", serde(bound(
         serialize = "T: Serialize",
         deserialize = "T: Deserialize<'de> + Copy + Default")))]
-    inner: [T; RAND_SIZE]
+    inner: [u32; RAND_SIZE]
 }
 
-impl<T> ::core::convert::AsRef<[T]> for IsaacArray<T> {
+impl IsaacArray32 {
     #[inline(always)]
-    fn as_ref(&self) -> &[T] {
-        &self.inner[..]
-    }
-}
-
-impl<T> ::core::ops::Deref for IsaacArray<T> {
-    type Target = [T; RAND_SIZE];
-    #[inline(always)]
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
-}
-
-impl<T> ::core::ops::DerefMut for IsaacArray<T> {
-    #[inline(always)]
-    fn deref_mut(&mut self) -> &mut [T; RAND_SIZE] {
+    pub fn inner_mut(&mut self) -> &mut [u32; RAND_SIZE] {
         &mut self.inner
     }
 }
 
-impl<T> ::core::default::Default for IsaacArray<T> where T: Copy + Default {
-    fn default() -> IsaacArray<T> {
-        IsaacArray { inner: [T::default(); RAND_SIZE] }
+impl AsRef<[u32]> for IsaacArray32 {
+    #[inline(always)]
+    fn as_ref(&self) -> &[u32] {
+        &self.inner[..]
+    }
+}
+
+impl Default for IsaacArray32 {
+    fn default() -> IsaacArray32 {
+        IsaacArray32 { inner: [0u32; RAND_SIZE] }
+    }
+}
+
+#[derive(Copy, Clone)]
+#[allow(missing_debug_implementations)]
+#[cfg_attr(feature="serde1", derive(Serialize, Deserialize))]
+pub struct IsaacArray64 {
+    #[cfg_attr(feature="serde1",serde(with="isaac_array_serde"))]
+    #[cfg_attr(feature="serde1", serde(bound(
+        serialize = "T: Serialize",
+        deserialize = "T: Deserialize<'de> + Copy + Default")))]
+    inner: [u64; RAND_SIZE]
+}
+
+impl IsaacArray64 {
+    pub fn inner_mut(&mut self) -> &mut [u64; RAND_SIZE] {
+        &mut self.inner
+    }
+}
+
+impl AsRef<[u32]> for IsaacArray64 {
+    #[inline(always)]
+    fn as_ref(&self) -> &[u32] {
+        unsafe {
+            let slice32 = slice::from_raw_parts(self.inner.as_ptr() as *const u32,
+                                                RAND_SIZE * 2);
+            slice32
+        }
+    }
+}
+
+impl Default for IsaacArray64 {
+    fn default() -> IsaacArray64 {
+        IsaacArray64 { inner: [0u64; RAND_SIZE] }
     }
 }