I don't think it is all that useful after all. A simple wrapper to implement `AsRef<[u32]>` for `[u64; 256]` is enough to make `Isaac64Rng` work.

Benchmarks before:
```
test gen_bytes_isaac64     ... bench:     386,986 ns/iter (+/- 1,188) = 2646 MB/s
test gen_u32_isaac64       ... bench:       3,091 ns/iter (+/- 13) = 1294 MB/s
test gen_u64_isaac64       ... bench:       3,757 ns/iter (+/- 3) = 2129 MB/s
```

After:
```
test gen_bytes_isaac64     ... bench:     385,762 ns/iter (+/- 231) = 2654 MB/s
test gen_u32_isaac64       ... bench:       2,168 ns/iter (+/- 5) = 1845 MB/s
test gen_u64_isaac64       ... bench:       4,139 ns/iter (+/- 3) = 1932 MB/s
```

So `next_u32` is now ~40% faster, `next_u64` ~10% slower. Seems acceptable to me, and a nice simplification.