@pitdicker 2018-04-12
> There's one problem I think: `Isaac64Rng` will now have different output on BE platforms.

Hmmm, I am starting to remember some troubles from a couple of months ago... Closing.