@vks 2018-04-12
Maybe add a comment to the code about the troubles?

On Thu, Apr 12, 2018, 13:12 Paul Dicker <notifications@github.com> wrote:

> There's one problem I think: Isaac64Rng will now have different output on
> BE platforms.
>
> Hmmm, I am starting to remember some troubles from a couple of months
> ago... Closing.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang-nursery/rand/pull/391#issuecomment-380768590>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACCtFj34zaeJ6uQsIut0CSGwUsCdrbPks5tnzaagaJpZM4TRLgQ>
> .
>