@pitdicker 2018-04-12
Good point, I'll try to write some comments.